+++ src/main.ts
@@ -29,6 +29,7 @@ import {
 import { setupLoginInvitationDialog } from "./utils/ui/loginInvitationDialog";
 import { setupNewLevelEventHandler } from "./levels/customEvents";
 import { setupMusic } from "./utils/audio";
+import { printVersion } from "./utils/ui/version";
 
 document.addEventListener("DOMContentLoaded", async () => {
   registerServiceWorker("cache");
@@ -61,4 +62,5 @@ document.addEventListener("DOMContentLoaded", async () => {
   setupScoresDialog();
   setupFullscreenBtn();
   setupMusic();
+  printVersion();
 });