+++ src/webComponents/links/TextLink.ts
@@ -10,6 +10,9 @@ class TextLink extends HTMLElement {
     const a = this.shadowRoot.querySelector("a");
     a.href = this.dataset.url;
     a.title = this.dataset.tooltip;
+    if (this.dataset.external === "true") {
+      a.target = "__blank";
+    }
   }
 }